+++ Elfie/Elfie/Model/Strings/UTF8.cs
@@ -8,9 +8,14 @@ namespace Microsoft.CodeAnalysis.Elfie.Model.Strings
         public const byte Null = (byte)'\0';
         public const byte Quote = (byte)'"';
         public const byte Comma = (byte)',';
+        public const byte Period = (byte)'.';
+        public const byte Backslash = (byte)'\\';
+        public const byte Pipe = (byte)'|';
         public const byte Tab = (byte)'\t';
         public const byte CR = (byte)'\r';
         public const byte LF = (byte)'\n';
+        public const byte Newline = (byte)'\n';
+        public const byte Space = (byte)' ';
         public const byte a = (byte)'a';
         public const byte z = (byte)'z';
         public const byte A = (byte)'A';
